Add Snapshot tests (#138)
* Add Snapshot tests
Snapshot test after 100 blocks.
In order for this to work, needed to track then "minimum height" that has been seen by a docker container, as Snapshots they will skip over the heights before the snapshot.

Also, several fixes for joins up/down, as well as a fix for some weird regex issues we were seeing in transaction queries on Github.