Invalid and Inactive Participant improvements (#407)
## Excluding Invalidated Participants entirely
### The Problem
Previous implementation had several key oversights when a participant was invalidated because of invalid inferences or downtime:
1. ActiveParticipants still included them, and gave no indication they had been excluded.
2. They were not removed from all EpochGroup sub-groups.

Additionally, the logic around removal was scattered and hard to understand, which is partly the reason these had not been addressed earlier.

### The Fix
#### ExcludedParticipants list
We add `ExcludedParticipants` to the state. It is keyed by epoch index and participant id. It includes the specific height of the exclusion and the reason for it. At the moment, exclusion is purely for invalidation because of too many invalid inferences. Other reasons are planned in the near future, thus the open-ended nature of the list.

When a participant is initially marked as invalidated for an epoch, they are added to the list.

#### ActiveParticipants response includes ExcludedParticipants
The response now includes the list of excluded participants, which clients will use to filter the Active list. The actual Active list remains the same, as it is important to have a cryptographically traceable ActiveParticipants list and this would become impractical if it also included arbitrary changes during an Epoch. (see related changes in this release).

#### Collateral has a hard dedupe for slashing
Slashing cannot rely on application logic to ensure that a single participant is slashed for a given reason only once per epoch. Given the dire nature of the punishment, we introduce a list in x/collateral that will ensure this at the lowest level, making application logic impossible to result in double punishment.

## Status update refactoring
### The Problem
1. Status update was not uniformly calculated when states changed, thus it would be easy to forget to call the right methods to check status. 
2. The algorithm for invalid status was naive and likely to have both false positives and false negatives, especially during the early phase of an Epoch
3. Inactivity was detected and punished only at the end of an epoch during Claims, leaving inactive participants on the network

### The Fix
#### Unified status calculation
- Status is now calculated alongside `SetParticipant`. Additionally, if a significant state change is detected(from ACTIVE to INVALID for instance), then the procedures for removing the participant are executed.
- Status calculations, for both Invalid and Inactive, have been standardized to use an SPRT algorithm. (see sprt.go and sprt.html in `calculations`.
- Once a participant is moved to INVALID or INACTIVE, it will not be moved back until the next epoch.
- Many governance parameters added to address this.

### Minor additional fixes
#### Participant status reset with new participant
A participant could "re-add" themselves and erase their current epoch stats for the epoch. This could both inadverantly punish good behavior by erasing it or allow masking of bad behavior.

#### Ramping status deprecated
Early plans for using ramping never materialized. For now, this status is superfluous and only adds complication and confusion.